[CI] Add Helm support for SSD-based stress tests and clean up charts (cta/CTA!933)
Closes #1197 and #1191

+++ continuousintegration/orchestration/helm/scheduler/scripts/reset_scheduler.sh
@@ -15,6 +15,8 @@
 #               granted to it by virtue of its status as an Intergovernmental Organization or
 #               submit itself to any jurisdiction.
 
+set -e
+
 echo "$(date '+%Y-%m-%d %H:%M:%S') [$(basename "${BASH_SOURCE[0]}")] Started"
 
 die() {
@@ -25,12 +27,16 @@ die() {
 echo "Using scheduler backend: $SCHEDULER_BACKEND"
 
 # Clean up scheduler
-if [ "$SCHEDULER_BACKEND" == "VFS" ]; then
+if [ "$SCHEDULER_BACKEND" == "vfs" ] || [ "$SCHEDULER_BACKEND" == "vfsDeprecated" ]; then
   echo "Installing the cta-objectstore-tools"
   dnf install -y cta-objectstore-tools
   echo "Wiping objectstore"
-  rm -fr $SCHEDULER_URL
-  mkdir -p $SCHEDULER_URL
+  if [ "$SCHEDULER_BACKEND" == "vfsDeprecated" ]; then
+    rm -fr $SCHEDULER_URL
+    mkdir -p $SCHEDULER_URL
+  else
+    rm -fr "${SCHEDULER_URL:?}/*"
+  fi
   cta-objectstore-initialize $SCHEDULER_URL || die "ERROR: Could not wipe the objectstore. cta-objectstore-initialize $SCHEDULER_URL FAILED"
   chmod -R 777 $SCHEDULER_URL
 elif [ "$SCHEDULER_BACKEND" == "postgres" ]; then
